refactor: 깊은 고민없이 패키지 분리한 부분 다시 원래대로 변경
