+++ src/main/java/ladder/Application.java
@@ -2,7 +2,7 @@ package ladder;
 
 import java.util.Scanner;
 import ladder.controller.LadderGameController;
-import ladder.domain.ladder.RandomBooleanGenerator;
+import ladder.domain.RandomBooleanGenerator;
 import ladder.view.InputView;
 import ladder.view.OutputView;
