+++ src/main/java/ladder/domain/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 public interface BooleanGenerator {
     boolean generate();
