+++ src/main/java/ladder/domain/Height.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 public class Height {
     private static final int HEIGHT_LOWER_BOUND = 1;
