+++ src/main/java/ladder/domain/LadderGame.java
@@ -1,12 +1,8 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import ladder.domain.item.Item;
-import ladder.domain.item.Items;
-import ladder.domain.player.Player;
-import ladder.domain.player.Players;
 
 public class LadderGame {
     private static final int SUBTRACT_LADDER_WIDTH_VALUE = 1;
