+++ src/main/java/ladder/domain/LineStatus.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 public enum LineStatus {
     CONNECTED,
