+++ src/main/java/ladder/domain/Position.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 import java.util.HashMap;
 import java.util.List;
