+++ src/test/java/ladder/domain/ItemNameTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.item;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
