+++ src/test/java/ladder/domain/ItemTest.java
@@ -1,8 +1,7 @@
-package ladder.domain.item;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import ladder.domain.ladder.Position;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
