+++ src/test/java/ladder/domain/ItemsTest.java
@@ -1,6 +1,6 @@
-package ladder.domain.item;
+package ladder.domain;
 
-import static ladder.domain.ladder.Position.valueOf;
+import static ladder.domain.Position.valueOf;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
