+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -1,11 +1,9 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
 import java.util.List;
-import ladder.domain.item.Items;
-import ladder.domain.player.Players;
 import ladder.util.TestBooleanGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
