+++ src/test/java/ladder/domain/LineStatusTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
