+++ src/test/java/ladder/domain/PlayerNameTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.player;
+package ladder.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
