+++ src/test/java/ladder/domain/PlayerTest.java
@@ -1,11 +1,8 @@
-package ladder.domain.player;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import ladder.domain.ladder.BooleanGenerator;
-import ladder.domain.ladder.Ladder;
-import ladder.domain.ladder.Position;
 import ladder.util.TestBooleanGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
