+++ src/test/java/ladder/domain/PlayersTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.player;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
@@ -9,9 +9,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import ladder.domain.ladder.BooleanGenerator;
-import ladder.domain.ladder.Ladder;
-import ladder.domain.ladder.Position;
 import ladder.util.TestBooleanGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
